requirements.txt
flask==3.0.0
flask-cors==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.7
python-dotenv==1.0.0
html5lib==1.1
soupsieve==2.5
